+++ src/01-one/conjugate/deconjugate.js
@@ -173,3 +173,5 @@ const deconjugate = function (str, isKnown) {
 }
 
 export default deconjugate
+// exported so tests/tagset.test.js can check these tags are declared
+export { suffixes }